+++ src/include/fdw.h
@@ -291,7 +291,6 @@ typedef struct ConnCacheEntry
 typedef enum
 {
 	CF_USUAL = 0,
-	CF_UNSHIPPABLE,				/* do not ship */
 	CF_SIGN_SUM,				/* SUM aggregation */
 	CF_SIGN_AVG,				/* AVG aggregation */
 	CF_SIGN_COUNT,				/* COUNT aggregation */